Solves issue #67 refactor(models): consolidate mixins into `models.mixins` - Move `HeaderMixin`, `CovarianceMixin`, and `VarianceMixin` into new `mosaicolabs.models.mixins`. - Remove legacy header_mixin.py and covariance_mixin.py. - Update imports across the SDK to `..mixins` (data, sensors, kinematics, dynamics, ROI, etc.). - Export mixins (and `BaseModel`) from `mosaicolabs.__init__` so public API remains stable. - Purely organizational refactor — no intended behavioral changes.
