+++ mosaico-sdk-py/src/mosaicolabs/models/__init__.py
@@ -1,6 +1,9 @@
 from .base_model import BaseModel as BaseModel
 from .header import Header as Header, Time as Time
 from .message import Message as Message
-from .covariance_mixin import CovarianceMixin as CovarianceMixin
-from .header_mixin import HeaderMixin as HeaderMixin
+from .mixins import (
+    HeaderMixin as HeaderMixin,
+    CovarianceMixin as CovarianceMixin,
+    VarianceMixin as VarianceMixin,
+)
 from .serializable import Serializable as Serializable
