+++ mosaico-sdk-py/src/mosaicolabs/models/sensors/pressure.py
@@ -6,7 +6,7 @@ Defines the data structure for pressure sensors.
 
 import pyarrow as pa
 
-from ..header_mixin import HeaderMixin
+from ..mixins import HeaderMixin
 from ..serializable import Serializable
